requirements.txt
alembic==1.0.9
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
gmail==0.6.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
Mako==1.0.9
MarkupSafe==1.1.1
marshmallow==2.19.0
marshmallow-enum==1.4.1
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.3
pyrsistent==0.14.11
pytest==4.3.1
pytest-flask==0.14.0
pytest-mock==1.10.2
python-dateutil==2.8.0
python-editor==1.0.4
redis==3.2.1
redlock-py==1.0.8
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
teamcity-messages==1.23
urllib3==1.24.1
webargs==5.1.3
Werkzeug==0.14.1